Run travis tests in PHP 5.2.17
As 5.2.17 is the last one of 5.2 branch, it would be ok to have this
tested as well despite of it isn't one of official supported version in
php.net nowadays.

In order to don't break completely our build, if it fails, 5.2.17 is
include in section `allow_failures` in Travis, keeping the build ok it
the newer definitions goes ok as well.

Signed-off-by: Rogerio Prado de Jesus <rogeriopradoj@gmail.com>