Clean extension repository before git pull
For at least memcached, if re-using the same repository to install a second version of PHP from the repository, it will fail to pull, and consequentially fail the build, due to not being able to properly pull the directory. It seems it creates/alters some files that are in source control and it cannot update.