Remove hard-patch code for APXS handling
Improves upon #204.

No longer any need for the manual patching of the Makefile as it can be done with sed and no longer applies to PHP7 (which always causes a Makefile.rej file to be created.)
This patch will put both the apache mod_php module in the phpenv folder rather than attempting to write it directly into the Apache folder, which may require root permission and is better to be managed separately for each PHP install.